requirements.txt
# DDR Generator — Python 3.10+
streamlit>=1.28.0
pymupdf>=1.23.0
openai>=1.40.0
python-docx>=1.1.0
reportlab>=4.0.0
Pillow>=10.0.0
# Optional: Google Gemini (set DDR_LLM_PROVIDER=gemini)
google-generativeai>=0.8.0
